refactor(api): finalize CORNETO 1.0 public surface
