+++ corneto/io/__init__.py
@@ -5,7 +5,6 @@ in various formats, including SIF and GML. It also includes functions to
 convert metabolic and signaling models to network graphs.
 """
 
-from ._graphio import load_corneto_graph, save_corneto_graph
 from ._metabolism import (
     cobra_model_to_graph,
     import_cobra_model,
@@ -18,9 +17,7 @@ __all__ = [
     "cobra_model_to_graph",
     "import_cobra_model",
     "import_miom_model",
-    "load_corneto_graph",
     "load_graph_from_sif",
     "load_graph_from_sif_tuples",
     "parse_cobra_model",
-    "save_corneto_graph",
 ]
